backend/user/urls.py
from django.urls import path, include
from user import views
from rest_framework.routers import DefaultRouter

app_name = 'user'
router = DefaultRouter()
router.register('', views.UserViewSet)

urlpatterns = [
    path('', include(router.urls)),
    path('token/', views.CreateTokenView.as_view(), name='token'),
    path('me/', views.ManageUserView.as_view(), name='me'),
]
