backend/requirements/base.txt
aioredis==1.3.1
amqp==2.6.1
aniso8601==7.0.0
argon2-cffi==19.1.0
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
autopep8==1.4.4
bcrypt==3.1.7
billiard==3.6.3.0
cachetools==4.2.0
celery==4.3.0
certifi==2019.9.11
cffi==1.12.3
channels==3.0.3
channels-redis==3.2.0
chardet==3.0.4
constantly==15.1.0
cryptography==3.3.1
daphne==3.0.1
Django==3.0
django-common-helpers==0.9.2
django-cors-headers==3.1.1
django-cron==0.5.1
django-crontab==0.7.1
django-discover-runner==1.0
django-enum-choices==2.1.1
django-filter==2.2.0
django-graphql-jwt==0.2.2
django-mandrill==0.1.0
django-redis==4.10.0
django-role-permissions==2.2.1
djangorestframework==3.10.3
docopt==0.4.0
facebook-sdk==3.1.0
fcm-django==0.3.7
google-api-core==1.24.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
graphene==2.1.8
graphene-django==2.6.0
graphql-core==2.2.1
graphql-relay==2.0.0
hiredis==1.1.0
html5lib==1.1
httplib2==0.18.1
hyperlink==21.0.0
idna==2.10
incremental==17.5.0
kombu==4.6.11
lxml==4.6.2
mandrill==1.0.59
msgpack==1.0.2
Pillow==6.1.0
promise==2.2.1
protobuf==3.14.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.19
pyfcm==1.4.9
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==20.0.1
PyPDF2==1.26.0
pytz==2019.2
redis==3.3.8
reportlab==3.5.56
requests==2.25.0
rsa==4.6
Rx==1.6.1
service-identity==18.1.0
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.3.0
Twisted==20.3.0
txaio==20.12.1
uritemplate==3.0.1
urllib3==1.26.2
vine==1.3.0
webencodings==0.5.1
xhtml2pdf==0.2.4
xlrd==1.2.0
zope.interface==5.2.0
